refactor: split main.py into focused modules (closes #603)
